refactor(transaction-history): move ActivityFeedCoordinator/Updater from :shared:activityfeed
Scaffolds :apps:flipcash:shared:transaction-history and relocates the two classes into it
(package com.flipcash.shared.transactionhistory). The old :shared:activityfeed keeps
@Deprecated forwarding typealiases and now depends on the new module via api() so callers
still importing com.flipcash.app.activityfeed resolve transitively. Zero dependent edits;
app assembles and the Hilt graph is intact. Plan Task 2.
