+++ apps/flipcash/shared/activityfeed/src/main/kotlin/com/flipcash/app/activityfeed/ActivityFeedUpdater.kt
@@ -1,18 +1,12 @@
 package com.flipcash.app.activityfeed
 
-import com.flipcash.app.core.updater.NetworkUpdater
-import javax.inject.Inject
-
 @Deprecated(
-    "Phasing out with the v2 (FeatureFlag.NewUi) launch: moving into :shared:transaction-history " +
-        "(com.flipcash.shared.transactionhistory) alongside ActivityFeedCoordinator. " +
+    "Moved to :shared:transaction-history as part of the v2 (FeatureFlag.NewUi) phase-out. " +
+        "Import com.flipcash.shared.transactionhistory.ActivityFeedUpdater instead. " +
         "Plan: docs/superpowers/plans/2026-08-11-activityfeed-phaseout.md",
+    ReplaceWith(
+        "ActivityFeedUpdater",
+        "com.flipcash.shared.transactionhistory.ActivityFeedUpdater",
+    ),
 )
-@Suppress("DEPRECATION") // constructor references the (also-deprecated) coordinator during phase-out
-class ActivityFeedUpdater @Inject constructor(
-    private val coordinator: ActivityFeedCoordinator,
-): NetworkUpdater() {
-    override suspend fun doUpdate() {
-        coordinator.fetchSinceLatest(count = 50)
-    }
-}
+typealias ActivityFeedUpdater = com.flipcash.shared.transactionhistory.ActivityFeedUpdater
