+++ apps/flipcash/shared/transaction-history/src/main/kotlin/com/flipcash/shared/transactionhistory/ActivityFeedUpdater.kt
@@ -1,0 +1,12 @@
+package com.flipcash.shared.transactionhistory
+
+import com.flipcash.app.core.updater.NetworkUpdater
+import javax.inject.Inject
+
+class ActivityFeedUpdater @Inject constructor(
+    private val coordinator: ActivityFeedCoordinator,
+): NetworkUpdater() {
+    override suspend fun doUpdate() {
+        coordinator.fetchSinceLatest(count = 50)
+    }
+}
